+semver:minor - Strong name the assemblies in this package (#1467)
* Strong name the assemblies in this package

* Don't strong name TUnit.Engine.Tests

This reduces the dependencies we need to get strong name signed down to just `EnumerableAsyncProcessor`.
Fortunately, no one in this repo uses InternalsVisibleTo to this project, or it would _have_ to be strong named.

* Don't strong name TUnit.Core.SourceGenerator.Tests

This avoids warnings about non-strong named TUnit assemblies that come from previously published nuget packages.
Once strong named assemblies are published to nuget.org and that newer version is consumed in this cycle, we *can* start strong naming this project if desired.

* Don't strong name test project

This eliminates the last two non-strong-named dependency warnings.

* Centralize policy to strong name shipping projects only

---------

Co-authored-by: Tom Longhurst <30480171+thomhurst@users.noreply.github.com>