Create tunit-ci-pipline.md
Added an example that took way to long for me to figure out and I stumbled on the answer by accident in the "discussions". So here's the information I was looking for. I also added some opinionated parts, but its never a bad thing to give some advice.

I had ChatGPT do editorial pass for spellchecking and linguistical flow, (I'm fluent in English, but not native, so I tend to construct clunky sentences)